requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptg==0.2.post4
cryptography==36.0.1
discord.py==2.0.0a4269+gb7e25645
fastapi==0.71.0
flanaapis
flanautils
frozenlist==1.2.0
greenlet==1.1.2
h11==0.12.0
hachoir==3.1.2
idna==3.3
iso8601==1.0.2
jellyfish==0.9.0
kaleido==0.2.1
mpmath==1.2.1
multibot
multidict==5.2.0
outcome==1.1.0
Pillow==9.0.0
playwright==1.17.2
plotly==5.5.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyee==8.2.2
pymongo==4.0.1
pyOpenSSL==21.0.0
requests==2.27.1
rsa==4.8
selenium==4.1.0
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.17.1
sympy==1.9
Telethon==1.24.0
tenacity==8.0.1
TikTokApi==4.1.0
trio==0.19.0
trio-websocket==0.9.2
twitchio==2.1.4
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.8
uvicorn==0.17.0
websockets==10.1
ws4py==0.5.1
wsproto==1.0.0
yarl==1.7.2
